+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * L2s and LQ CTMs with missing input files can be made anyway after a certain number of days in https://github.com/punch-mission/punchpipe/pull/194
 * Reduces L2 code duplication in https://github.com/punch-mission/punchpipe/pull/195
 * Set date_obs correctly in DB for L2s, and update cleaner flow for L2 in https://github.com/punch-mission/punchpipe/pull/196
+* Temporarily disabled NFI in L2s in https://github.com/punch-mission/punchpipe/pull/197
 
 ## Version 0.0.9: June 4, 2025